lint options, importing provenance-lib-core works, ProvVis doesnt
