+++ tsconfig.json
@@ -3,7 +3,7 @@
     "declaration": true,
     "noImplicitAny": true,
     "noEmitOnError": true,
-    "noUnusedLocals": true,
+    "noUnusedLocals": false,
     "esModuleInterop": true,
     "preserveWatchOutput": true,
     "module": "commonjs",
