+++ tslint.json
@@ -103,7 +103,7 @@
         ],
         "quotemark": [
             true,
-            "single",
+            "double",
             "avoid-escape"
         ],
         "radix": true,
